Toggle like with button
There is now a button to like a song as a user.
This button will be disabled if there is no song currently being
played.
This button will be a thumbs up if the song is not yet liked, and when
clicked it will like the song.
This button will be a thumbs down if the song is already liked, and
when clicked it will remove the relationship between a user and a song.
If a song is already liked when loading the button will start in the
thumbs down. If the button is pressed, then the button will change to
the opposite type.